requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
pymongo==4.5.0
requests==2.31.0
certifi==2023.7.22
Flask-Cors==3.1.1